Rewrote to pass distance as constructor
